pyproject.toml
[project]
name = "uniswap-v3-multiagent"
version = "1.0.0"
description = "Uniswap V3 Multi-Agent Risk Analysis System with LangGraph, A2A, and Fast MCP"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Andrea Bagno"}
]

dependencies = [
    # Core Framework
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Web Frameworks
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "starlette>=0.41.0",
    
    # HTTP Clients
    "httpx>=0.28.0",
    "requests>=2.32.0",
    
    # Data Processing
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "scipy>=1.14.0",
    
    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # A2A Protocol
    "a2a-sdk",
    
    # Fast MCP
    "fastmcp>=0.1.0",
    
    # LangChain MCP Adapters (for MCP tool calling)
    "langchain-mcp-adapters>=0.2.0",
    
    # Search & Intelligence
    "tavily-python>=0.5.0",
    
    # Streamlit UI
    "streamlit>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common_ai", "backend", "pool_risk_service", "token_intel_service"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
